+++ Unreal/CarlaUnreal/Plugins/Carla/Source/Carla/Server/CarlaServer.cpp
@@ -55,6 +55,7 @@
 #include <carla/rpc/VehiclePhysicsControl.h>
 #include <carla/rpc/VehicleLightState.h>
 #include <carla/rpc/VehicleLightStateList.h>
+#include <carla/rpc/VehicleTelemetryData.h>
 #include <carla/rpc/WalkerBoneControlIn.h>
 #include <carla/rpc/WalkerBoneControlOut.h>
 #include <carla/rpc/WalkerControl.h>
@@ -1910,6 +1911,31 @@ BIND_SYNC(is_sensor_enabled_for_ros) << [this](carla::streaming::detail::stream_
           " Actor Id: " + FString::FromInt(ActorId));
     }
     return R<void>::Success();
+  };
+
+  BIND_SYNC(get_telemetry_data) << [this](
+     cr::ActorId ActorId) -> R<cr::VehicleTelemetryData>
+  {
+    REQUIRE_CARLA_EPISODE();
+    FCarlaActor* CarlaActor = Episode->FindCarlaActor(ActorId);
+    if (!CarlaActor)
+    {
+      return RespondError(
+          "get_telemetry_data",
+          ECarlaServerResponse::ActorNotFound,
+          " Actor Id: " + FString::FromInt(ActorId));
+    }
+    FVehicleTelemetryData TelemetryData;
+    ECarlaServerResponse Response =
+        CarlaActor->GetVehicleTelemetryData(TelemetryData);
+    if (Response != ECarlaServerResponse::Success)
+    {
+      return RespondError(
+          "get_telemetry_data",
+          Response,
+          " Actor Id: " + FString::FromInt(ActorId));
+    }
+    return cr::VehicleTelemetryData(TelemetryData);
   };
 
   BIND_SYNC(show_vehicle_debug_telemetry) << [this](